1. Fixed recycler out of memory bug. 2. Changed cache time. 3. Small refactoring.

+++ app/src/main/java/ru/nbdev/mediadownloader/model/network/pixabay/PixabayApiService.java
@@ -29,6 +29,6 @@ public interface PixabayApiService {
     @GET("api")
     Single<PixabayApiPhotosList> getPhotoById(
             @Query("key") String key,
-            @Query("id") int id
+            @Query("id") long id
     );
 }
